moved api_server stuff out to distinct test file
